sync

+++ macro/recoPdf.C
@@ -43,7 +43,7 @@ void recoPdf(TString in="hits.root", TString pdf="hits.pdf.root", double timeres
   }
 
   double time;
-  int nph[5]={0};
+  double nph[5]={0};
   int tnph(0),totalf(0),totals(0), ch(0), pdg(0);
   
   for (int ievent=0; ievent<10000 && ievent<prt_entries; ievent++){
